fix: units double

+++ app/src/main/java/com/bulletapps/candypricer/data/parameters/CreateProductParameters.kt
@@ -8,7 +8,7 @@ data class CreateProductParameters(
     @SerializedName("unit_id")
     var unitId: Int,
     @SerializedName("quantity")
-    var quantity: Int,
+    var quantity: Double,
     @SerializedName("supplies")
     var suppliesId: List<Int>,
     @SerializedName("profit_margin")
