+++ app/src/main/java/com/bulletapps/candypricer/data/response/ProductResponse.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class ProductResponse(
     @SerializedName("name") val name: String,
     @SerializedName("unit") val unit: UnitResponse?,
-    @SerializedName("quantity") val quantity: Int,
+    @SerializedName("quantity") val quantity: Double,
     @SerializedName("profit_margin") val profitMargin: Double?,
     @SerializedName("labor_work") val laborValue: Double?,
     @SerializedName("variable_expenses") val variableExpenses: Double?,
