+++ app/src/main/java/com/bulletapps/candypricer/data/response/SupplyResponse.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class SupplyResponse(
     @SerializedName("id") val id: Int,
     @SerializedName("name") val name: String,
-    @SerializedName("quantity") val quantity: Int,
+    @SerializedName("quantity") val quantity: Double,
     @SerializedName("value") val value: Double,
     @SerializedName("unit") val unit: UnitResponse?
 )
